fix glide_test.go
